requirements.txt
Flask
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
Werkzeug==1.0.1
numpy
pandas
tensorflow-cpu